movie workie bitty

+++ src/networking/protocol/network_position.rs
@@ -5,12 +5,12 @@ use naia_shared::{Property, Replicate};
 #[derive(Component, Replicate)]
 #[protocol_path = "crate::networking::protocol::Protocol"]
 pub struct NetworkPosition {
-    pub x: Property<i16>,
-    pub y: Property<i16>,
+    pub x: Property<f32>,
+    pub y: Property<f32>,
 }
 
 impl NetworkPosition {
-    pub fn new(x: i16, y: i16) -> Self {
+    pub fn new(x: f32, y: f32) -> Self {
         NetworkPosition::new_complete(x, y)
     }
 }
